fix(risk-engine): resolve full external audit — packaging, state protocol, calibration, parity
P0 packaging:
- risk-php bundles the canonical Lua at resources/risk-v1.lua (self-contained;
  monorepo realpath removed); bundle's post-autoload-dump hack deleted
- kiwicaptcha-risk includes ../resources/risk-v1.lua; Cargo dep
  kiwicaptcha = { version = "1.1.0", path = "../kiwicaptcha" }
- all three Lua copies are sha256-identical (CI asserts)

P1 state protocol:
- session + principal state are now updated/saved with event-specific
  semantics (AuthSuccess/ProtectedActionSuccess/ConfirmedLegitimate →
  trust; AuthFailure/ProtectedActionFailure/ConfirmedAbuse → pressure);
  principal_credit is real
- epoch-correct pseudonyms: prev/current/next ids each HMAC'd with their
  own epoch (sourceIdForEpoch/subnetIdForEpoch); ±1 keys are observer-only
- ONLY PreIssue counts as a request; feedback events mutate only their own
  channels; assess() vs record_feedback() (no limiter, no decision);
  record() deprecated alias
- all 14 event semantics implemented (ExpiredChallenge/ProtectedActionSuccess/
  AuthenticationSuccess/AuthenticationFailure/RateLimitHit)
- shared duplicate semantics: duplicate event_id returns current signals
  with is_duplicate=1 (no error, no zero vector); idempotency_key usable
  through the public API; EventReceipt exposed
- scope u32 everywhere

P1 policy:
- policy parser invariants (version equality, base_risk 0..1000, scope
  1..=u32::MAX, global floors exactly 5 with index 0 = Allow)
- Argon capacity checked LAST → floors can never reintroduce Argon
  (capacity < 300 → StepUp)
- degraded mode applies global_floors[last_known_level]
- top-4 positive contributor reasons (override reasons first, cap 4)
- unknown scopes: reject (controller issues default profile) or minimum
  (sha20/sha20) mode

P1 calibration:
- bounded Redis aggregate buckets (scope × hour × band × action,
  48h TTL, ≤24 keys/scope), JSON-string decision receipts (EX 300,
  GETDEL); bias applied in BOTH languages with identical integer math

P1 performance:
- Rust Redis store: round-robin connection pool (4 default)
- network classifier: bitwise radix trie (O(prefix depth)) in both
  languages, identical OR-of-matching-prefixes semantics
- Rust global limiter aligned to the in-process fixed-window model (PHP
  parity), fail-open Redis gone

P1 Symfony wiring:
- ResourcePressureProviderInterface + RedisRiskHealthProvider (semaphore
  slots, issuance usage, Redis health); global_per_second admission cap
- global_pressure.enabled / argon_capacity.enabled operational
- post_solve_check enforced in the validator (post-solve deny → violation)
- unknown_scope config (reject|minimum)
- single samesite (strict) / http_only nodes (duplicates removed)
- audited Argon profiles: 16/32/64 MiB, t=3, p=1, target bits 1; the
  Argon→Sha20 mapping removed (SHA deployments can still issue Argon)
- session TTL (continuity cookie) reaches the store

P2 worker: kiwicaptacha-wasm.js typo fixed (4 occurrences)

+++ packages/kiwicaptcha-risk/src/context.rs
@@ -8,7 +8,7 @@ use crate::resources::ResourcePressure;
 
 /// Inputs of one risk assessment.
 pub struct RiskContext<'a> {
-    pub scope: u16,
+    pub scope: u32,
     pub source_ip: IpAddr,
     /// Raw session cookie value (pseudonymized before storage).
     pub session_id: Option<&'a [u8]>,
@@ -22,7 +22,7 @@ pub struct RiskContext<'a> {
 impl<'a> RiskContext<'a> {
     /// Convenience constructor for tests and simple call sites.
     pub fn new(
-        scope: u16,
+        scope: u32,
         source_ip: IpAddr,
         session_id: Option<&'a [u8]>,
         principal_id: Option<&'a [u8]>,
